refactor: Put models in separate files
